Fix heartbeat overflow
sseHeartbeatIntervalMillis is an Uint32, hence it is wrong to assume its
value can safely fit in an int.

Use .toJava() instead, which returns a proper long, exposing the fact
that ServerSseHandler is using the wrong type.

+++ transport/transport-http/src/main/java/org/opendaylight/netconf/transport/http/SseUtils.java
@@ -48,7 +48,7 @@ public final class SseUtils {
      *      no ping message will be sent by server
      */
     public static void enableServerSse(final Channel channel, final EventStreamService service,
-            final int maxFieldValueLength, final int heartbeatIntervalMillis) {
+            final int maxFieldValueLength, final long heartbeatIntervalMillis) {
         final var sseHandler =
             new ServerSseHandler(requireNonNull(service), maxFieldValueLength, heartbeatIntervalMillis);
         requireNonNull(channel).pipeline().addLast(new ChannelInboundHandlerAdapter() {
